add check for active reservation

+++ rooms-booking-frontend/src/service/UserService.js
@@ -35,10 +35,14 @@ export function deleteAccount() {
   getAxios()
     .delete(apiURL + userId)
     .then((response) => {
-      localStorage.clear();
-      toast.info(response.data);
-      window.setTimeout(() => window.location.replace('http://localhost:3000/'), 3000);
-      
+      if(response.data === 'Account deleted successfully') {
+        localStorage.clear();
+        toast.info(response.data);
+        window.setTimeout(() => window.location.replace('http://localhost:3000/'), 3000);
+      }
+      else {
+        toast.info(response.data);
+      }
     })
     .catch((error) => {
       console.log(error);
